+++ components/org.wso2.transport.http.netty/src/test/java/org/wso2/transport/http/netty/util/server/initializers/EchoServerInitializer.java
@@ -59,7 +59,7 @@ public class EchoServerInitializer extends HTTPServerInitializer {
         private HttpResponse httpResponse;
         private int responseStatusCode = 200;
         private boolean chunked = true;
-        private int contentLength = 0;
+        private long contentLength = 0;
         private boolean keepAlive = false;
         private BlockingQueue<HttpContent> content = new LinkedBlockingQueue<>();
